Fixed comments from PR
