+++ integration-tests/trait-dyn-cross-contract-calls/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "incrementer-e2e-test"
+name = "trait-incrementer-caller"
 version = "4.0.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2021"
@@ -12,11 +12,10 @@ scale = { package = "parity-scale-codec", version = "3", default-features = fals
 scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
 
 dyn-traits = { path = "./traits", default-features = false }
-trait-incrementer = { path = "./contracts/incrementer", default-features = false, features = ["ink-as-dependency"] }
-trait-incrementer-caller = { path = "./contracts/incrementer-caller", default-features = false, features = ["ink-as-dependency"] }
 
 [dev-dependencies]
 ink_e2e = { path = "../../crates/e2e" }
+trait-incrementer = { path = "./contracts/incrementer", default-features = false, features = ["ink-as-dependency"] }
 
 [lib]
 path = "lib.rs"
@@ -29,8 +28,6 @@ std = [
     "scale-info/std",
 
     "dyn-traits/std",
-    "trait-incrementer/std",
-    "trait-incrementer-caller/std",
 ]
 e2e-tests = []
 ink-as-dependency = []
@@ -39,5 +36,4 @@ ink-as-dependency = []
 [workspace]
 members = [
     "contracts/incrementer",
-    "contracts/incrementer-caller",
 ]
