--- integration-tests/trait-dyn-cross-contract-calls/contracts/incrementer-caller/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "trait-incrementer-caller"
-version = "4.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-publish = false
-
-[dependencies]
-ink = { path = "../../../../crates/ink", default-features = false }
-
-scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
-scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
-
-dyn-traits = { path = "../../traits", default-features = false }
-
-[lib]
-path = "lib.rs"
-
-[features]
-default = ["std"]
-std = [
-    "ink/std",
-    "scale/std",
-    "scale-info/std",
-    "dyn-traits/std",
-]
-ink-as-dependency = []
