--- integration-tests/trait-dyn-cross-contract-calls/contracts/incrementer-caller/lib.rs
@@ -1,36 +1,0 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-#![allow(clippy::new_without_default)]
-
-#[ink::contract]
-pub mod caller {
-    use dyn_traits::Increment;
-
-    /// The caller of the incrementer smart contract.
-    #[ink(storage)]
-    pub struct Caller {
-        /// Here we accept a type which implements the `Incrementer` ink! trait.
-        incrementer: ink::contract_ref!(Increment),
-    }
-
-    impl Caller {
-        /// Creates a new caller smart contract around the `incrementer` account id.
-        #[ink(constructor)]
-        pub fn new(incrementer: AccountId) -> Self {
-            Self {
-                incrementer: incrementer.into(),
-            }
-        }
-    }
-
-    impl Increment for Caller {
-        #[ink(message)]
-        fn inc(&mut self) {
-            self.incrementer.inc()
-        }
-
-        #[ink(message)]
-        fn get(&self) -> u64 {
-            self.incrementer.get()
-        }
-    }
-}
